smartstudy/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.25.0
pymongo>=4.5.0
motor>=3.3.0
python-dotenv>=1.0.0
beanie>=1.20.0
PyPDF2>=3.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0